добавил Readme.md,  изменил название тестов,. разбил Responses на файлы, исправил вывод ошибок
В Readme.md написал как получить Api ключ и как добавить его в код ;
написал названия для тестов, отражая их назначение ;
Response для постов, и для информации о группе, разбил на 1 файл = 1 класс ;
ошибки о доступах к группе или постам со стены паблика, теперь выводятся не в консоль, а как исключение;
написал комменты к методам, описывающие, что делает метод.
