+++ TaktTusur.Media.Clients.VkApi/GroupInfoResponse/GroupByIdResponse.cs
@@ -1,0 +1,12 @@
+﻿using Elasticsearch.Net;
+
+namespace TaktTusur.Media.Clients.VkApi.GroupInfoResponse;
+
+public class GroupByIdResponse
+{
+    public Response Response { get; set; }
+    public Error GroupInfoError { get; set; }
+}
+
+
+
