+++ TaktTusur.Media.Clients.VkApi/VkApiOptions.cs
@@ -3,5 +3,4 @@
 public class VkApiOptions
 {
     public string Key { get; set; }
-
 }
