+++ TaktTusur.Media.Clients.VkApi/WallByIdRequest.cs
@@ -1,9 +1,4 @@
 ﻿using RestSharp;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace TaktTusur.Media.Clients.VkApi
 {
