+++ TaktTusur.Media.Clients.VkApi/WallByIdResponse/Doc.cs
@@ -1,0 +1,37 @@
+﻿using System.Text.Json.Serialization;
+
+namespace TaktTusur.Media.Clients.VkApi.WallByIdResponse
+{
+    public class Doc
+    {
+        [JsonPropertyName("id")]
+        public int Id { get; set; }
+
+        [JsonPropertyName("owner_id")]
+        public int OwnerId { get; set; }
+
+        [JsonPropertyName("title")]
+        public string Title { get; set; }
+
+        [JsonPropertyName("size")]
+        public int Size { get; set; }
+
+        [JsonPropertyName("ext")]
+        public string Ext { get; set; }
+
+        [JsonPropertyName("date")]
+        public int Date { get; set; }
+
+        [JsonPropertyName("type")]
+        public int Type { get; set; }
+
+        [JsonPropertyName("url")]
+        public string Url { get; set; }
+
+        [JsonPropertyName("is_unsafe")]
+        public int IsUnsafe { get; set; }
+
+        [JsonPropertyName("access_key")]
+        public string AccessKey { get; set; }
+    }
+}
