+++ TaktTusur.Media.Clients.VkApi/WallByIdResponse/Donut.cs
@@ -1,0 +1,10 @@
+﻿using System.Text.Json.Serialization;
+
+namespace TaktTusur.Media.Clients.VkApi.WallByIdResponse
+{
+    public class Donut
+    {
+        [JsonPropertyName("is_donut")]
+        public bool IsDonut { get; set; }
+    }
+}
