+++ TaktTusur.Media.Clients.VkApi/WallByIdResponse/Image.cs
@@ -1,0 +1,19 @@
+﻿using System.Text.Json.Serialization;
+
+namespace TaktTusur.Media.Clients.VkApi.WallByIdResponse
+{
+    public class Image
+    {
+        [JsonPropertyName("url")]
+        public string Url { get; set; }
+
+        [JsonPropertyName("width")]
+        public int Width { get; set; }
+
+        [JsonPropertyName("height")]
+        public int Height { get; set; }
+
+        [JsonPropertyName("with_padding")]
+        public int WithPadding { get; set; }
+    }
+}
