+++ TaktTusur.Media.Clients.VkApi/WallByIdResponse/PostSource.cs
@@ -1,0 +1,10 @@
+﻿using System.Text.Json.Serialization;
+
+namespace TaktTusur.Media.Clients.VkApi.WallByIdResponse
+{
+    public class PostSource
+    {
+        [JsonPropertyName("type")]
+        public string Type { get; set; }
+    }
+}
