+++ TaktTusur.Media.Clients.VkApi/WallByIdResponse/Size.cs
@@ -1,0 +1,19 @@
+﻿using System.Text.Json.Serialization;
+
+namespace TaktTusur.Media.Clients.VkApi.WallByIdResponse
+{
+    public class Size
+    {
+        [JsonPropertyName("height")]
+        public int Height { get; set; }
+
+        [JsonPropertyName("type")]
+        public string Type { get; set; }
+
+        [JsonPropertyName("width")]
+        public int Width { get; set; }
+
+        [JsonPropertyName("url")]
+        public string Url { get; set; }
+    }
+}
