+++ TaktTusur.Media.Clients.VkApi/WallByIdResponse/Views.cs
@@ -1,0 +1,10 @@
+﻿using System.Text.Json.Serialization;
+
+namespace TaktTusur.Media.Clients.VkApi.WallByIdResponse
+{
+    public class Views
+    {
+        [JsonPropertyName("count")]
+        public int Count { get; set; }
+    }
+}
