+++ TaktTusur.Media.Clients.VkApi/WallByIdResponse/WallPostByIdResponse.cs
@@ -1,0 +1,15 @@
+﻿using Elasticsearch.Net;
+using System.Text.Json.Serialization;
+
+namespace TaktTusur.Media.Clients.VkApi.WallByIdResponse
+{
+    public class WallPostByIdResponse
+    {
+        [JsonPropertyName("response")]
+        public WallResponse Response { get; set; }
+
+        [JsonPropertyName("error")]
+        public Error Error { get; set; }
+    }
+
+}
