+++ TaktTusur.Media.Clients.VkApi/WallByIdResponse/WallResponse.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Text.Json.Serialization;
+using System.Threading.Tasks;
+
+namespace TaktTusur.Media.Clients.VkApi.WallByIdResponse
+{
+    public class WallResponse
+    {
+        [JsonPropertyName("count")]
+        public int Count { get; set; }
+
+        [JsonPropertyName("items")]
+        public List<Item> Items { get; set; }
+    }
+}
